# docs/swagger.yaml
basePath: /api
definitions:
  controllers.AccountInputJSON:
    properties:
      bank_name:
        type: string
      card_number:
        type: string
      card_type:
        description: Ви просили CardType
        type: string
      color:
        type: string
      currency:
        type: string
      goal_id:
        type: string
      initial_balance:
        type: integer
      name:
        type: string
      payment_system:
        type: string
      storage_type_id:
        type: string
      type:
        type: string
      user_id:
        type: string
    required:
    - currency
    - name
    - type
    type: object
...
  models.Account:
    properties:
      balance:
        description: Поточний кешований баланс
        type: integer
      bank_name:
        type: string
      card_number:
        description: Банківські деталі
        type: string
      card_type:
        description: Credit/Debit
        type: string
      color:
        type: string
      created_at:
        type: integer
      currency:
        type: string
      deleted_at:
        description: Індекс для soft delete
        type: integer
      family_id:
        type: string
      goal:
        $ref: '#/definitions/models.Goal'
      goal_id:
        description: Зв'язок з ціллю
        type: string
      id:
        type: string
      initial_balance:
        description: Баланси
        type: integer
      is_archived:
        type: boolean
      is_group:
        type: boolean
      is_synced:
        type: boolean
      name:
        type: string
      parent_id:
        description: Для вкладеності
        type: string
      payment_system:
        description: Visa/Mastercard
        type: string
      storage_type:
        $ref: '#/definitions/models.StorageType'
      storage_type_id:
        description: Для скарбничок (Piggy Bank)
        type: string
      type:
        description: 'Типи: ''card'', ''cash'', ''crypto'', ''piggy_bank'''
        type: string
      updated_at:
        type: integer
      user_id:
        type: string
    type: object
...
  models.Goal:
    properties:
      accounts:
        description: Relations
        items:
          $ref: '#/definitions/models.Account'
        type: array
      color:
        type: string
      created_at:
        type: integer
      currency:
        type: string
      current_amount:
        description: Вираховується динамічно
        type: integer
      date_deadline:
        type: integer
      date_start:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      description:
        type: string
      external_link:
        type: string
      family_id:
        type: string
      hidden_from:
        description: ID юзера, від якого приховано
        type: string
      icon:
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      name:
        type: string
      photo_url:
        type: string
      remove_photo:
        type: boolean
      status:
        description: active, reached, failed
        type: string
      target_amount:
        type: integer
      updated_at:
        type: integer
      user_id:
        description: '<-- ДОДАНО: Власник цілі'
        type: string
      visibility:
        description: '<-- ДОДАНО: Налаштування приватності'
        type: string
    type: object
...
  models.StorageType:
    properties:
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      family_id:
        description: Pointer для NULL (системні типи)
        type: string
      icon:
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      is_system:
        type: boolean
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: integer
    type: object
...
info:
  contact:
    name: Vlad Hrytsaiuk
  description: API for WeGaS Finance application.
  title: WeGaS Finance API
  version: "1.0"
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: Returns a list of all accounts accessible to the current user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Account'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all accounts
      tags:
      - Accounts
    post:
      consumes:
      - application/json
      description: Creates a new financial account for the current user.
      parameters:
      - description: Account details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.AccountInputJSON'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new account
      tags:
      - Accounts
  /accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an account by its ID.
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete account
      tags:
      - Accounts
    get:
      consumes:
      - application/json
      description: Returns a single account by its ID.
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get account by ID
      tags:
      - Accounts
    put:
      consumes:
      - application/json
      description: Updates an existing account by its ID.
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated account details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.AccountInputJSON'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update account
      tags:
      - Accounts
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"